fix first set of comments
