--- .gitignore
@@ -1,4 +1,0 @@
-src/build/
-src/skai.egg-info/
-src/skai/model/__pycache__/
-src/skai/model/configs/__pycache__/
