+++ src/skai/model/configs/skai_two_tower_config.py
@@ -4,6 +4,7 @@ r"""Configuration file for experiment with SKAI data and TwoTower model.
 
 
 import ml_collections
+
 from skai.model.configs import base_config
